[listener] add twitter split
